requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
fonttools==4.61.1
frozenlist==1.8.0
googletrans==4.0.2
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
kiwisolver==1.4.9
magic-filter==1.0.12
matplotlib==3.10.8
multidict==6.7.0
numpy==2.2.6
packaging==25.0
pillow==12.1.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
